+++ lubee/src/@common/types/EmojisDataTypes.ts
@@ -21,13 +21,6 @@ export interface TitleBoxProps {
   subtitleText?: string;
 }
 
-export interface CommentModalProps {
-  handleCloseBtn: () => void;
-  profileIconSrc: string;
-  commentText: string;
-  setCommentText?: (text: string) => void;
-}
-
 export interface FinalEmojisDataTypes {
   account: string;
   emojiData: EmojisDataTypes[];
