+++ lubee/src/home/today/types/CommentModalTypes.ts
@@ -1,0 +1,6 @@
+export interface CommentModalProps {
+  handleCloseBtn: () => void;
+  profileIconSrc: string;
+  commentText: string;
+  setCommentText?: (text: string) => void;
+}
