--- lubee/src/layout/OnboardingLayout.tsx
@@ -1,5 +1,0 @@
-import { Outlet } from "react-router-dom";
-
-export default function OnboardingLayout() {
-  return <Outlet />;
-}
