+++ lubee/src/onboarding/custom/index.tsx
@@ -7,6 +7,7 @@ import ProgressBar from "../components/ProgressBar";
 import OnboardingTitleBox from "../components/OnboardingTitleBox";
 import OnboardingBtn from "../components/OnboardingBtn";
 import { profileCustomIconsData } from "@common/core/ProfileCustomIconsData";
+
 interface CustomProps {
   moveToOnboardingCode: () => void;
   moveToOnboardingProfile: () => void;
